reorganize Atoms in state folder
