+++ src/component/LikeButton.tsx
@@ -1,12 +1,12 @@
 import { HeartOutlineIcon, HeartSolidIcon } from './Icons'
 import toast from 'react-hot-toast'
 import { atom, Atom, PrimitiveAtom, useAtom, useAtomValue, useSetAtom } from 'jotai'
-import { nowPlayingIsLikedAtom, nowPlayingTrackIdAtom } from 'state/Atoms'
 import { useEffect, useMemo } from 'react'
 import { useQueryClient } from '@tanstack/react-query'
 import { useRemoveSavedTracksMutation, useSaveTracksMutation, useTrackLikeQuery } from './sdk/ClientHooks'
 import { useTransientAtom } from 'platform/hook/useTransientAtom'
 import { classNames } from 'util/Utils'
+import { nowPlayingIsLikedAtom, nowPlayingTrackIdAtom } from 'state/NowPlayingAtom'
 
 
 interface LikeButtonProps {
