+++ src/component/detailedReview/table/MemoHeader.tsx
@@ -5,9 +5,9 @@
 
 import { atom, useSetAtom } from 'jotai'
 import { memo } from 'react'
-import { selectedTrackAtom } from 'state/Atoms'
 import { ReviewGroupHeader, ReviewGroupHeaderProps } from './GroupHeader'
 import { expandedGroupsAtom } from './TableAtoms'
+import { selectedTrackAtom } from 'state/SelectedTrackAtom'
 
 type MemoHeaderProps = Omit<ReviewGroupHeaderProps, 'onClick'>
