+++ src/state/SelectedTrackAtom.ts
@@ -1,0 +1,4 @@
+import { atom } from 'jotai'
+
+export interface SelectedTrack { reviewId: string, trackId: string }
+export const selectedTrackAtom = atom<SelectedTrack | undefined>(undefined)
